refactor: move heartbeat runs list into modal with icon-based controls
Replace the inline Heartbeat Runs card with three icon buttons in the
checklist card header: play (run now), clock (history modal), and gear
(config modal). The runs list is now shown in a dedicated HeartbeatRunsView
sheet modal, consistent with HeartbeatConfigView.
